* Removed old texture creation files * Moved Model folder to client package
