+++ src/main/java/com/sigmundgranaas/forgero/client/forgerotool/model/ToolModel2DManager.java
@@ -1,8 +1,8 @@
-package com.sigmundgranaas.forgero.item.forgerotool.model;
+package com.sigmundgranaas.forgero.client.forgerotool.model;
 
 import com.sigmundgranaas.forgero.Forgero;
-import com.sigmundgranaas.forgero.item.forgerotool.model.dynamicmodel.Dynamic2DModelFactory;
-import com.sigmundgranaas.forgero.item.forgerotool.model.dynamicmodel.DynamicModel;
+import com.sigmundgranaas.forgero.client.forgerotool.model.dynamicmodel.Dynamic2DModelFactory;
+import com.sigmundgranaas.forgero.client.forgerotool.model.dynamicmodel.DynamicModel;
 import com.sigmundgranaas.forgero.item.forgerotool.tool.item.ForgeroTool;
 import com.sigmundgranaas.forgero.item.forgerotool.toolpart.ForgeroToolPartItem;
 import net.fabricmc.fabric.api.renderer.v1.model.FabricBakedModel;
