+++ src/main/java/com/sigmundgranaas/forgero/client/forgerotool/model/ToolPartModelType.java
@@ -1,4 +1,4 @@
-package com.sigmundgranaas.forgero.item.forgerotool.model;
+package com.sigmundgranaas.forgero.client.forgerotool.model;
 
 import java.util.Locale;
