+++ src/main/java/com/sigmundgranaas/forgero/client/forgerotool/model/dynamicmodel/DynamicModel.java
@@ -1,4 +1,4 @@
-package com.sigmundgranaas.forgero.item.forgerotool.model.dynamicmodel;
+package com.sigmundgranaas.forgero.client.forgerotool.model.dynamicmodel;
 
 import net.minecraft.client.render.model.UnbakedModel;
 import net.minecraft.client.render.model.json.JsonUnbakedModel;
