+++ src/main/java/com/sigmundgranaas/forgero/client/forgerotool/model/dynamicmodel/binding/PickaxeBindingModel.java
@@ -1,10 +1,10 @@
-package com.sigmundgranaas.forgero.item.forgerotool.model.dynamicmodel.binding;
+package com.sigmundgranaas.forgero.client.forgerotool.model.dynamicmodel.binding;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import com.sigmundgranaas.forgero.Forgero;
-import com.sigmundgranaas.forgero.item.forgerotool.model.dynamicmodel.AbstractDynamicModel;
+import com.sigmundgranaas.forgero.client.forgerotool.model.dynamicmodel.AbstractDynamicModel;
 import com.sigmundgranaas.forgero.item.forgerotool.toolpart.ForgeroToolPartItem;
 import net.minecraft.client.util.ModelIdentifier;
