+++ src/main/java/com/sigmundgranaas/forgero/client/forgerotool/model/dynamicmodel/head/ShovelHeadModel.java
@@ -1,9 +1,9 @@
-package com.sigmundgranaas.forgero.item.forgerotool.model.dynamicmodel.head;
+package com.sigmundgranaas.forgero.client.forgerotool.model.dynamicmodel.head;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
-import com.sigmundgranaas.forgero.item.forgerotool.model.dynamicmodel.AbstractDynamicModel;
+import com.sigmundgranaas.forgero.client.forgerotool.model.dynamicmodel.AbstractDynamicModel;
 import com.sigmundgranaas.forgero.item.forgerotool.toolpart.ForgeroToolPartItem;
 
 public class ShovelHeadModel extends AbstractDynamicModel {
