+++ src/main/java/com/sigmundgranaas/forgero/client/forgerotool/texture/ForgeroTextureIdentifier.java
@@ -1,6 +1,6 @@
 package com.sigmundgranaas.forgero.client.forgerotool.texture;
 
-import com.sigmundgranaas.forgero.item.forgerotool.model.ToolPartModelType;
+import com.sigmundgranaas.forgero.client.forgerotool.model.ToolPartModelType;
 import com.sigmundgranaas.forgero.item.forgerotool.toolpart.ForgeroToolPartTypes;
 
 import java.util.Optional;
