+++ src/main/java/com/sigmundgranaas/forgero/client/forgerotool/texture/ForgeroTextureIdentifierImpl.java
@@ -1,7 +1,7 @@
 package com.sigmundgranaas.forgero.client.forgerotool.texture;
 
+import com.sigmundgranaas.forgero.client.forgerotool.model.ToolPartModelType;
 import com.sigmundgranaas.forgero.item.forgerotool.Constants;
-import com.sigmundgranaas.forgero.item.forgerotool.model.ToolPartModelType;
 import com.sigmundgranaas.forgero.item.forgerotool.toolpart.ForgeroToolPartItem;
 import com.sigmundgranaas.forgero.item.forgerotool.toolpart.ForgeroToolPartTypes;
 import com.sigmundgranaas.forgero.utils.FileSystemUtilities;
