+++ src/main/java/com/sigmundgranaas/forgero/client/forgerotool/texture/template/TemplateTexture.java
@@ -1,7 +1,7 @@
 package com.sigmundgranaas.forgero.client.forgerotool.texture.template;
 
+import com.sigmundgranaas.forgero.client.forgerotool.texture.PixelInformation;
 import com.sigmundgranaas.forgero.client.forgerotool.texture.material.MaterialPalette;
-import com.sigmundgranaas.forgero.item.forgerotool.model.PixelInformation;
 
 import java.awt.image.BufferedImage;
 import java.util.ArrayList;
